check phone number

+++ src/main/java/com/controller/NotepadController.java
@@ -1,5 +1,6 @@
 package com.controller;
 
+import com.checkData.CheckData;
 import com.model.Person;
 import com.service.NotepadService;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -8,12 +9,14 @@ import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
+
 import javax.servlet.http.HttpServletRequest;
 
 
 @Controller
 @RequestMapping("/")
 public class NotepadController {
+    private String errorMessage=null;
 
     @Autowired
     private NotepadService notepadService;
@@ -22,16 +25,21 @@ public class NotepadController {
     public String getNotepadList(Model model) {
         model.addAttribute("notepadList", notepadService.notepadList());
         model.addAttribute("notepadListModify", notepadService.notepadListModify());
+        model.addAttribute("errorPhone", errorMessage);
         return "notepad";
     }
 
     @RequestMapping(value = "modify", method = RequestMethod.GET)
     public String serviceStudent(@ModelAttribute Person person, HttpServletRequest req) {
-        if (req.getParameter("add") != null) {
-            notepadService.addPersonToNotepadList(person);
-        }
-        if (req.getParameter("del") != null) {
-            notepadService.deletePersonFromNotepadList(person);
+        if (CheckData.checkPhone(person) == 0) {
+            if (req.getParameter("add") != null) {
+                notepadService.addPersonToNotepadList(person);
+            }
+            if (req.getParameter("del") != null) {
+                notepadService.deletePersonFromNotepadList(person);
+            }
+        } else {
+           errorMessage="Phone number is wrong!";
         }
         return "redirect:/";
     }
